pyproject.toml
[tool.poetry]
package-mode = false
name = "django_store"
version = "0.1.0"
description = "Simple Demo of Django"


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-simple-captcha = "^0.6.0"
django-tinymce = "^4.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-mptt = "^0.16.0"
pytils = "^0.4.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
